requirements.txt
anyio==4.9.0
beautifulsoup4==4.13.3
bs4==0.0.2
certifi==2025.1.31
charset-normalizer==3.4.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
lxml==5.3.2
numpy==2.2.4
pandas==2.2.3
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-telegram-bot==22.0
pytz==2025.2
redis==5.2.1
requests==2.32.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
typing-extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
